requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
sentence-transformers==2.2.2
tavily-python==0.3.3
langgraph==0.0.62
langchain==0.1.0
langchain-google-genai==1.0.0
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0